requirements.txt
python-telegram-bot>=20.0
SQLAlchemy>=1.4
psycopg2-binary
APScheduler>=3.9